Added fasttext example
