+++ .github/workflows/test-notebooks.yml
@@ -45,6 +45,7 @@ jobs:
           - product-tours/how-it-works/docs/Neptune-API-Tour.ipynb
           - quick-starts/first-run/docs/Use-Neptune-API-to-log-your-first-run.ipynb
           - use-cases/time-series-forecasting/walmart-sales/notebooks/EDA.ipynb
+          - use-cases/nlp/classification/fasttext/code/fasttext_nb.ipynb
         exclude:
           - notebooks: integrations-and-supported-tools/sacred/notebooks/Neptune_Sacred.ipynb
             python-version: "3.7"
