+++ .github/workflows/test-scripts.yml
@@ -44,6 +44,7 @@ jobs:
           - integrations-and-supported-tools/transformers/scripts
           - integrations-and-supported-tools/xgboost/scripts
           - use-cases/time-series-forecasting/walmart-sales/scripts
+          - use-cases/nlp/classification/fasttext/code
         exclude:
           - scripts: integrations-and-supported-tools/sacred/scripts
             python-version: "3.7"
