+++ use-cases/nlp/classification/fasttext/code/run_examples.sh
@@ -1,0 +1,7 @@
+set -e
+
+echo "Installing requirements..."
+pip install -U -r requirements.txt
+
+echo "Running fasttext_script.py..."
+python fasttext_script.py
