+++ use-cases/nlp/classification/keras/code/requirements.txt
@@ -1,5 +1,5 @@
+neptune
 neptune-tensorflow-keras
-neptune_client
-numpy==1.23.5
-pydot==1.4.2
-tensorflow==2.11.0
+numpy
+pydot
+tensorflow
