+++ use-cases/nlp/summarization/hf_transformers/requirements.txt
@@ -1,11 +1,11 @@
 accelerate
-datasets >= 1.8.0
+datasets
 evaluate
-neptune-client
+neptune
 nltk
 protobuf
 py7zr
 rouge-score
-sentencepiece != 0.1.92
-torch >= 1.3
-transformers == 4.22.0
+sentencepiece
+torch
+transformers
